Remove the wildcard /login/* route in favor of more scoped ones
The edX integration needs /login/_private/complete to work (which is an OAuth2 route) - since we have /login/* APISIX eats this and some scheduled tasks fail, notably the repair faulty users one.

I added /login/oidc because we were using it on the frontend so not 100% sure it's strictly necessary.