feat: Add support for ValKey engine in Elasticache (#3280)
* feat: Add support for ValKey engine in Elasticache

The ValKey engine is a fork of Redis that remains API compatible. It has added support
for better multi-thread/multi-process functionality so that it can be faster. It is als
a minimum of 20% cheaper than Redis in Elasticache. This adds logic to support ValKey as
an engine, and sets that as the default. This also updates edX environments to use
ValKey 7.2 by default. The upgrade from Redis to ValKey is non-destructive.

* config: Swap to Valkey engine