fix: Properly handle Traefik middleware naming
In the gateway API implementation for the ExtensionRef filter type Traefik will automatically prepend the namespace to the name of the middleware when it looks it up. This leads to a mismatch in the name that the middleware uses and the name that the gateway is trying to resolve. This updates the logic to address that disparity.