config,fix: Use proper peer address for Concourse web nodes
The Concourse web nodes need to have a peer address that is resolvable by the other web
nodes. The Consul address _can_ resolve to the web node but it is a random chance of if
the right node will get contacted. This updates the peer address to use the specific
node IP of the instance so that there is a guarantee of getting the right instance every time.