fix: Set placeholder to fix broken build on master branch
The openassessment xblock uses a `hasattr` lookup for the `LMS_ROOT_URL` and falls back
to an empty string when that attribute isn't present. The resent configuration
restructuring n edx-platform adds an assignment of `LMS_ROOT_URL = None` as a default
value which causes the logic
[here](https://github.com/openedx/edx-ora2/blob/master/openassessment/xblock/load_static.py#L19)
to break due to trying to call a method on the newly discovered `None` value.