config: Move studio and preview dmains behind courses domain
As we consolidate more services under the "learn" domain, the semantics of those domains
become more important, as does the preservation of semantics. This moves the studio and
preview domains for the edX deployment to be rooted under the "courses.learn" domain
used by the LMS to signify their semantic relationship to the courses in question,
rather than other arbitrary studio and preview functions in the "Learn" context.