config: Enable LTI provider functionality in Learn edX (#3523)
In order to more smoothly manage the Canvas bot functionality we want to allow the Learn
edX to be an LTI provider so that it doesn't conflict with Canvas usage that is
dependent residential MITx. In the Canvas case we don't need to associate the user
accounts with their Kerberos, so there's not real benefit from the MITx residential edX
being the provider. This allows us to have an LTI provider setup that is customized for
the Canvas bot use case without conflicting with the residential use case.