config: Move MITx Online edX domains to be rooted at learn (#3496)
* config: Move MITx Online edX domains to be rooted at learn

* config: Update edxapp platform name

* fix: Set proper DNS zone for notes domain changes

* config: Move studio and preview dmains behind courses domain

As we consolidate more services under the "learn" domain, the semantics of those domains
become more important, as does the preservation of semantics. This moves the studio and
preview domains for the edX deployment to be rooted under the "courses.learn" domain
used by the LMS to signify their semantic relationship to the courses in question,
rather than other arbitrary studio and preview functions in the "Learn" context.

* refactor: Update to take a mapping of source to target domains

This adds stacks for CI and QA to allow for progressive rollout of redirects for MITx
Online edX and any future staged redirects. This also simplifies the VCL to be a bit
more aintainable.

* fix: Properly manage Fastly TLS subscription configuration ID

* feat: Manage TLS certs in Fastly for edxapp

* config: Domain target fixes

* feat: Add pipeline for Fastly Redirector

* config: Codify organization toggle for paltform realm

* config: Don't use learn URLs for production apps yet

The release cadence of learn-ai, mitxonline, and micromasters doesn't match the cadence
of edx-platform releases. This reverts the domain changes for production so that we can
manage that rollout more explicitly and not block other releases.