feat: Add optional explicit mapping for SAML attributes
The NTUA SSO has a documented set of SAML attributes that can be released. This updates
the IDP mappings to use those documented attributes and map them into the proper
Keycloak attributes. This also allows for seting the mapping of Keycloak attributes to
SAML friendly names in the SAML config.