config: Don't override host value for Airbyte API responses
We already have logic to handle pagination in the API via Dagster which is currently our
only reliance on the Airbyte API, so this isn't necessary. It also causes a weird bug in
the pagination logic that adds the host as the first portion of the path in the request.