hack: Add wildcard to allowed hosts due to ELB health checks
The ELB health checks set a host header of the private IP of the EC2 instances. With the
removal of the wildcard in the config, and actually using the host values that we
specified, the healthchecks are now all erroring. This is a temporary fix until we are
in Kubernetes and can do a better job of controlling that configuration.