config: Remove consul-esm from Consul deployment
We are drawing down our usage of Consul, and the Consul ESM was never actively relied
upon. Its purpose was to do health checking of external services like RDS which we would
have other signals for failure to resolve. This component is also spewing a huge amount
of useless error messages into our Grafana logs