feat: Add Fastly management to MITx Online project (#3785)
The login flow was landing on the API domain instead of the frontend domain when going
through fastly. This was because of the host name override being set which caused the
login link to point at the API domain instead of the frontend domain. This brings the
Fastly management under Pulumi control, as well as setting up APISix to handle the
frontend domain for routing as well so that there isn't a need to route directly to the
API domain for login/logout.