config: Remove expiry of Keycloak user events
We often need to have a full picture of account activity so we don't want to expire
events after only one day. Until we start experiencing issues with data volumes, we
should keep the expiration turned off. Even if we turn it back on, it needs to be
significantly longer than one day. Presumably on the order of 90 days.