fix: Remove invalid Output type membership check in edxapp k8s_resources
The code was attempting to use 'in' operator on a Pulumi Output object
which is not iterable. This caused: TypeError: argument of type 'Output'
is not iterable.

Simplified to directly use applications_vpc for k8s deployments, since
edxapp k8s clusters are always deployed in applications_vpc (not the
target_vpc which is for EC2 deployments).